+++ Newton/src/functions/vnr_internalenergy_evolution.c
@@ -18,7 +18,7 @@ void internal_energy_evolution_VNR(void *variables, double *newton_var, size_t s
     allocVecForOMP(size_of_pb, 0., &dummy);
     allocVecForOMP(size_of_pb, 0., &specific_volume);
 #pragma omp parallel for
-    for (int i=0; i<size_of_pb; ++i) {
+    for (size_t i=0; i<size_of_pb; ++i) {
         specific_volume[i] = 1. / vars->density_new[i];
 #ifdef DEBUG
         printf("specific_volume[%zu] = %15.9g | density_new[%zu] = %15.9g\n", i, specific_volume[i], i, vars->density_new[i]);
@@ -28,7 +28,7 @@ void internal_energy_evolution_VNR(void *variables, double *newton_var, size_t s
     vars->miegruneisen->solve(vars->miegruneisen, size_of_pb, specific_volume, newton_var, pression, dpsurde, dummy);
     double delta_v = 0.;
 #pragma omp parallel for
-    for (int i=0; i<size_of_pb; ++i) {
+    for (size_t i=0; i<size_of_pb; ++i) {
         delta_v = 1. / vars->density_new[i] - 1. / vars->density_old[i];
         // Fonction à annuler
         func[i] = newton_var[i] + pression[i] * delta_v / 2. + vars->pressure[i] * delta_v / 2. - vars->internal_energy_old[i];
